Moved functions into *Route.js file Deleted instanceMetrics.js
