+++ src/server/cpuRoute.js
@@ -1,14 +1,49 @@
-var express = require('express');
- 
-var cpu = new express.Router();
 
-var metrics = require('./instanceMetrics')
+var params = {
+    EndTime: 1431440584,
+    /* required */
+    MetricName: 'CPUUtilization',
+    /* required */
+    Namespace: 'AWS/EC2',
+    /* required */
+    Period: 3600,
+    /* required */
+    StartTime: 1431440584 - 3600,
+    /* required */
+    Statistics: ['Average'],
+    Dimensions: [{
+            Name: 'InstanceId',
+            /* required */
+            Value: 'i-192650ef' /* required */
+        },
+        /* more items */
+    ],
+    Unit: 'Percent'
+};
 
+//Will this require instance metrics as well to compute 
+module.exports = function(req, res){
+	var instanceCollection
+	for ( var i in instances){
+		instanceCollection = instances.Reservations[r].Instances[i];
+		console.log(instanceCollection);
+	}
+    console.log("req: " + req.query.value);
+    var val = req.query.value;
+    var rEndTime = parseInt(req.query.endTime);
+    var rStartTime = parseInt(req.query.startTime);
+    var rMetricName = req.query.metric;
 
-cpu.get('/', function(req, res, next) {
-	console.log("CHECKIDY");
-    res.send(metrics.getCPU(req, res));
-});
- 
+    params.Dimensions[0].Value = val;
+    params.EndTime = rEndTime;
+    params.StartTime = rStartTime;
+    params.MetricName = rMetricName;
+    params.Unit = 'Percent';
 
-module.exports = cpu;
+    var cloudwatch = new AWS.CloudWatch();
+    cloudwatch.getMetricStatistics(params, function(err, data) {
+    	// console.log(data)
+        res.send(data);
+    });
+};
+
